exynos: cpufreq: Reset cpuinfo_transition_latency to default
cpuinfo_transition_latency -The time it takes on this CPU to switch between two frequencies in nano seconds. If unknown or known to be that high that the driver does not work with the ondemand governor, -1 (CPUFREQ_ETERNAL) will be returned. Using this information can be useful to choose an appropriate polling frequency for a kernel governor or userspace daemon. Make sure to not switch the frequency too often resulting in performance loss.